aca-project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
aiofiles==23.2.1
numpy==1.26.4
opencv-python-headless==4.10.0.82
torch==2.3.1
torchvision==0.18.1
Pillow==10.3.0
